POC Mantine - refactor and tidy, enable onClicks
